Fixing array out of bounds exception.
git-svn-id: svn+ssh://external.hltcoe.jhu.edu/home/hltcoe/mgormley/public/repos/dep_parse_filtered/trunk@699 0408881f-cb4b-4750-8b71-c94bcb194269

+++ src/no/uib/cipr/matrix/sparse/longs/FastSparseLVector.java
@@ -32,7 +32,7 @@ public class FastSparseLVector extends SparseLVector {
     double norm2Value;
 
     public FastSparseLVector() {
-        super(Integer.MAX_VALUE);
+        super(Long.MAX_VALUE);
     }
 
     public FastSparseLVector(long size) {
@@ -48,11 +48,11 @@ public class FastSparseLVector extends SparseLVector {
     }
 
     public FastSparseLVector(long[] index, double[] data) {
-        this(Integer.MAX_VALUE, index, data);
+        this(Long.MAX_VALUE, index, data);
     }
 
     public FastSparseLVector(double[] denseData) {
-        super(Integer.MAX_VALUE, LongSort.getIndexArray(denseData), denseData);
+        super(Long.MAX_VALUE, LongSort.getIndexArray(denseData), denseData);
     }
 
     /**
